// src/types/supplier/supplier.ts
export interface Supplier {
  id: string;
  id_atos?: string;
  name: string;
  comercial_name?: string;
  link_web?: string;
  created_at_atos?: string;
  assignee_atos?: string;
  created_by_atos?: string;
  full_address?: string;
  zip?: string;
  state?: string;
  presentation?: string;
  size?: string;
  iso_9001_2015?: boolean;
  iatf?: boolean;
  enabled: boolean;
  capacity?: string;
  hs_company_id?: string;
  contact_assigned?: string;
  description?: string;
  phone?: string;
  image?: string;
  first_contact?: string;
  token_bid?: string;
  type?: string;
} 